refactor(step02): replace wrong step03 to 02
