+++ src/simple-todos/step02/.meteor/.finished-upgraders
@@ -17,3 +17,18 @@ notices-for-facebook-graph-api-2
 1.5-add-dynamic-import-package
 1.7-split-underscore-from-meteor-base
 1.8.3-split-jquery-from-blaze
+notices-for-0.9.0
+notices-for-0.9.1
+0.9.4-platform-file
+notices-for-facebook-graph-api-2
+1.2.0-standard-minifiers-package
+1.2.0-meteor-platform-split
+1.2.0-cordova-changes
+1.2.0-breaking-changes
+1.3.0-split-minifiers-package
+1.4.0-remove-old-dev-bundle-link
+1.4.1-add-shell-server-package
+1.4.3-split-account-service-packages
+1.5-add-dynamic-import-package
+1.7-split-underscore-from-meteor-base
+1.8.3-split-jquery-from-blaze
