+++ src/simple-todos/step02/.meteor/.id
@@ -4,4 +4,4 @@
 #   - ensuring you don't accidentally deploy one app on top of another
 #   - providing package authors with aggregated statistics
 
-gu733n85xrzb.z56vudnb2r4
+ndsl3ducl0p.poz2g1vkys0r
