+++ src/simple-todos/step02/.meteor/packages
@@ -4,22 +4,19 @@
 # 'meteor add' and 'meteor remove' will edit this file for you,
 # but you can also edit it by hand.
 
-meteor-base@1.4.0             # Packages every Meteor app needs to have
+meteor-base@1.5.1             # Packages every Meteor app needs to have
 mobile-experience@1.1.0       # Packages for a great mobile UX
-mongo@1.10.1                   # The database Meteor supports right now
-reactive-var@1.0.11            # Reactive variable for tracker
+mongo@1.16.6                   # The database Meteor supports right now
+reactive-var@1.0.12            # Reactive variable for tracker
 
-standard-minifier-css@1.7.2   # CSS minifier run for production mode
-standard-minifier-js@2.6.0    # JS minifier run for production mode
+standard-minifier-css@1.9.2   # CSS minifier run for production mode
+standard-minifier-js@2.8.1    # JS minifier run for production mode
 es5-shim@4.8.0                # ECMAScript 5 compatibility for older browsers
-ecmascript@0.15.0              # Enable ECMAScript2015+ syntax in app code
-typescript@4.1.2              # Enable TypeScript syntax in .ts and .tsx modules
+ecmascript@0.16.7              # Enable ECMAScript2015+ syntax in app code
+typescript@4.9.4              # Enable TypeScript syntax in .ts and .tsx modules
 shell-server@0.5.0            # Server-side component of the `meteor shell` command
+hot-module-replacement@0.5.3  # Update client in development without reloading the page
 
-tracker@1.2.0                 # Dependency tracker to allow reactive callbacks
-static-html             # Define static page content in .html files
-akryum:vue-component    # Vue-CLI template to publish components
-
-meteortesting:mocha               # A package for writing and running your meteor app and package tests with mocha
-johanbrook:publication-collector  # Test a Meteor publication by collecting its output
+static-html@1.3.2             # Define static page content in .html files
+vite:bundler
 autopublish
