+++ src/simple-todos/step02/.meteor/release
@@ -1,1 +1,1 @@
-METEOR@2.1
+METEOR@2.12
