+++ src/simple-todos/step02/.prettierrc
