+++ src/simple-todos/step02/client/main.js
@@ -1,16 +1,1 @@
-import Vue from 'vue'
-
-import '../imports/ui/plugins'
-
-import App from '../imports/ui/App.vue'
-
-import VueMeteorTracker from 'vue-meteor-tracker';
-
-Vue.use(VueMeteorTracker);
-
-Meteor.startup(() => {
-  new Vue({
-    el: '#app',
-    ...App,
-  })
-})
+// main entry point is in imports/ui/main.jsx
