+++ src/simple-todos/step02/server/main.js
@@ -1,5 +1,5 @@
-import { Meteor } from 'meteor/meteor';
-import { TasksCollection } from '/imports/api/TasksCollection';
+import { Meteor } from 'meteor/meteor'
+import { TasksCollection } from '../imports/api/TasksCollection'
 
 const insertTask = taskText => TasksCollection.insert({ text: taskText });
 
@@ -12,7 +12,7 @@ Meteor.startup(() => {
       'Fourth Task',
       'Fifth Task',
       'Sixth Task',
-      'Seventh Task',
-    ].forEach(insertTask);
+      'Seventh Task'
+    ].forEach(insertTask)
   }
 });
