+++ src/simple-todos/step02/tailwind.config.js
