+++ src/simple-todos/step02/tests/main.js
@@ -1,21 +1,20 @@
-import { Meteor } from 'meteor/meteor';
-import assert from 'assert';
+import assert from 'assert'
 
-describe('simple-todos-vue', function() {
-  it('package.json has correct name', async function() {
-    const { name } = await import('../package.json');
-    assert.strictEqual(name, 'simple-todos-vue');
-  });
+describe('vue-skeleton', function () {
+  it('package.json has correct name', async function () {
+    const { name } = await import('../package.json')
+    assert.strictEqual(name, 'vue-skeleton')
+  })
 
   if (Meteor.isClient) {
-    it('client is not server', function() {
-      assert.strictEqual(Meteor.isServer, false);
-    });
+    it('client is not server', function () {
+      assert.strictEqual(Meteor.isServer, false)
+    })
   }
 
   if (Meteor.isServer) {
-    it('server is not client', function() {
-      assert.strictEqual(Meteor.isClient, false);
-    });
+    it('server is not client', function () {
+      assert.strictEqual(Meteor.isClient, false)
+    })
   }
-});
+})
