+++ src/simple-todos/step03/.meteor/.id
@@ -4,4 +4,4 @@
 #   - ensuring you don't accidentally deploy one app on top of another
 #   - providing package authors with aggregated statistics
 
-ndsl3ducl0p.poz2g1vkys0r
+gu733n85xrzb.z56vudnb2r4
