+++ src/simple-todos/step03/.meteor/release
@@ -1,1 +1,1 @@
-METEOR@2.12
+METEOR@2.1
