+++ src/simple-todos/step03/.meteor/versions
@@ -1,72 +1,81 @@
-allow-deny@1.1.1
+akryum:vue-component@0.15.2
+akryum:vue-component-dev-client@0.4.7
+akryum:vue-component-dev-server@0.1.4
+allow-deny@1.1.0
 autopublish@1.0.7
-autoupdate@1.8.0
-babel-compiler@7.10.4
-babel-runtime@1.5.1
+autoupdate@1.7.0
+babel-compiler@7.6.0
+babel-runtime@1.5.0
 base64@1.0.12
 binary-heap@1.0.11
-blaze-tools@1.1.3
+blaze-tools@1.0.10
 boilerplate-generator@1.7.1
 caching-compiler@1.2.2
-caching-html-compiler@1.2.1
-callback-hook@1.5.1
-check@1.3.2
-ddp@1.4.1
-ddp-client@2.6.1
+caching-html-compiler@1.1.3
+callback-hook@1.3.0
+check@1.3.1
+ddp@1.4.0
+ddp-client@2.4.0
 ddp-common@1.4.0
-ddp-server@2.6.1
-diff-sequence@1.1.2
-dynamic-import@0.7.3
-ecmascript@0.16.7
-ecmascript-runtime@0.8.1
-ecmascript-runtime-client@0.12.1
-ecmascript-runtime-server@0.11.0
-ejson@1.1.3
+ddp-server@2.3.2
+deps@1.0.12
+diff-sequence@1.1.1
+dynamic-import@0.6.0
+ecmascript@0.15.0
+ecmascript-runtime@0.7.0
+ecmascript-runtime-client@0.11.0
+ecmascript-runtime-server@0.10.0
+ejson@1.1.1
 es5-shim@4.8.0
-fetch@0.1.3
-geojson-utils@1.0.11
+fetch@0.1.1
+geojson-utils@1.0.10
 hot-code-push@1.0.4
-hot-module-replacement@0.5.3
-html-tools@1.1.3
-htmljs@1.1.1
-id-map@1.1.1
+html-tools@1.0.11
+htmljs@1.0.11
+http@1.0.10
+id-map@1.1.0
+insecure@1.0.7
 inter-process-messaging@0.1.1
-launch-screen@1.3.0
-logging@1.3.2
-meteor@1.11.2
-meteor-base@1.5.1
-minifier-css@1.6.4
-minifier-js@2.7.5
-minimongo@1.9.3
+johanbrook:publication-collector@1.1.0
+launch-screen@1.2.0
+livedata@1.0.18
+logging@1.2.0
+meteor@1.9.3
+meteor-base@1.4.0
+meteortesting:browser-tests@1.3.4
+meteortesting:mocha@2.0.1
+meteortesting:mocha-core@8.0.1
+minifier-css@1.5.3
+minifier-js@2.6.0
+minimongo@1.6.1
 mobile-experience@1.1.0
 mobile-status-bar@1.1.0
-modern-browsers@0.1.9
-modules@0.19.0
-modules-runtime@0.13.1
-modules-runtime-hot@0.14.2
-mongo@1.16.6
-mongo-decimal@0.1.3
+modern-browsers@0.1.5
+modules@0.16.0
+modules-runtime@0.12.0
+mongo@1.10.1
+mongo-decimal@0.1.2
 mongo-dev-server@1.1.0
-mongo-id@1.0.8
-npm-mongo@4.16.0
+mongo-id@1.0.7
+npm-mongo@3.8.1
 ordered-dict@1.1.0
-promise@0.12.2
-random@1.2.1
-react-fast-refresh@0.2.7
-reactive-var@1.0.12
+promise@0.11.2
+random@1.2.0
+react-fast-refresh@0.1.0
+reactive-var@1.0.11
 reload@1.3.1
 retry@1.1.0
-routepolicy@1.1.1
+routepolicy@1.1.0
 shell-server@0.5.0
-socket-stream-client@0.5.1
-spacebars-compiler@1.3.1
-standard-minifier-css@1.9.2
-standard-minifier-js@2.8.1
-static-html@1.3.2
-templating-tools@1.2.2
-tracker@1.3.2
-typescript@4.9.4
-underscore@1.0.13
-vite:bundler@0.1.9
-webapp@1.13.5
-webapp-hashing@1.1.1
+socket-stream-client@0.3.1
+spacebars-compiler@1.1.3
+standard-minifier-css@1.7.2
+standard-minifier-js@2.6.0
+static-html@1.2.2
+templating-tools@1.1.2
+tracker@1.2.0
+typescript@4.1.2
+underscore@1.0.10
+url@1.3.1
+webapp@1.10.0
+webapp-hashing@1.1.0
