+++ src/simple-todos/step03/client/main.css
@@ -1,3 +1,3 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
+.new-task {
+    margin-top: 1rem;
+}
