+++ src/simple-todos/step03/package.json
@@ -1,19 +1,17 @@
 {
-  "name": "step01",
+  "name": "simple-todos-vue",
   "private": true,
   "scripts": {
     "start": "meteor run",
-    "build": "meteor build ../output/vue --directory",
     "test": "meteor test --once --driver-package meteortesting:mocha",
     "test-app": "TEST_WATCH=1 meteor test --full-app --driver-package meteortesting:mocha",
     "visualize": "meteor --production --extra-packages bundle-visualizer"
   },
   "dependencies": {
-    "@babel/runtime": "^7.22.6",
-    "meteor-node-stubs": "^1.2.1",
-    "vue": "^3.2.45",
-    "vue-meteor-tracker": "^3.0.0-beta.7",
-    "vue-router": "^4.1.6"
+    "@babel/runtime": "^7.11.2",
+    "meteor-node-stubs": "^1.0.1",
+    "vue": "^2.6.12",
+    "vue-meteor-tracker": "^2.0.0-beta.5"
   },
   "meteor": {
     "mainModule": {
@@ -21,19 +19,5 @@
       "server": "server/main.js"
     },
     "testModule": "tests/main.js"
-  },
-  "devDependencies": {
-    "@prettier/plugin-php": "^0.19.6",
-    "@types/meteor": "^2.8.1",
-    "@vitejs/plugin-vue": "^3.2.0",
-    "autoprefixer": "^10.4.13",
-    "eslint": "^8.44.0",
-    "eslint-config-prettier": "^8.8.0",
-    "eslint-plugin-prettier": "^5.0.0",
-    "eslint-plugin-vue": "^9.15.1",
-    "postcss": "^8.4.19",
-    "prettier": "^3.0.0",
-    "tailwindcss": "^3.2.4",
-    "vite": "^3.2.3"
   }
 }
