+++ src/simple-todos/step03/tests/main.js
@@ -1,20 +1,21 @@
-import assert from 'assert'
+import { Meteor } from 'meteor/meteor';
+import assert from 'assert';
 
-describe('vue-skeleton', function () {
-  it('package.json has correct name', async function () {
-    const { name } = await import('../package.json')
-    assert.strictEqual(name, 'vue-skeleton')
-  })
+describe('simple-todos-vue', function() {
+  it('package.json has correct name', async function() {
+    const { name } = await import('../package.json');
+    assert.strictEqual(name, 'simple-todos-vue');
+  });
 
   if (Meteor.isClient) {
-    it('client is not server', function () {
-      assert.strictEqual(Meteor.isServer, false)
-    })
+    it('client is not server', function() {
+      assert.strictEqual(Meteor.isServer, false);
+    });
   }
 
   if (Meteor.isServer) {
-    it('server is not client', function () {
-      assert.strictEqual(Meteor.isClient, false)
-    })
+    it('server is not client', function() {
+      assert.strictEqual(Meteor.isClient, false);
+    });
   }
-})
+});
